requirements.txt
anyio==3.6.2
app==0.0.1
asynctest==0.13.0
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
deprecation==2.1.0
exceptiongroup==1.1.1
fastapi==0.96.0
greenlet==2.0.2
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pydantic==1.10.7
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.26.1
tomli==2.0.1
trinsic-sdk==1.11.0
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
